Auth Signing sort of working
