+++ .gitignore
@@ -1,3 +1,5 @@
 dumps
 *.pyc
 venv
+libencrypt.so
+encrypt.dll
